requirements.txt
biopython==1.72
certifi==2018.8.24
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.0
numpy==1.14.6
pandas==0.22.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
Werkzeug==1.0.1
